requirements.txt
python-telegram-bot>=21.7
sqlalchemy>=2.0.30
python-dotenv>=1.0.0